Updated GUI
The Cards are generated using the CardBank.deck though does not show multiples of a single card per player. Left Side Game State Status Map implemented though currently it just shows smaller versions of the same cards. Buttons were also modified to be highlighted green or dark red depending on the current player's options. On a related note, the Buttons array needs to be attributed to each player as this can be used to show a player their current moves. It would make the game very user friendly