Inherited game logic from Main
Manually copied & pasted the files from Rallas's 11/20 commit. This includes the updates in Actions, App, GameState & Person and the prototype. The only changes implemented was the update of the prototype GUI (it should handle the vectors of hands now & deal w/ multiples of a 1 kind of card). Note: Compiles fine on my end though does not recognize game states atm