GUI is Functional Again
The Card graphic generation sequence has been fixed though the cards are redrawn each time a GameState is received. I changed the Surrender button -> Clear Cards Button as I'm still exploring a solution for this as implementing the function call at the end of the card generation sequence causes the clear function to act constantly. There's also a HARDCODED max cap for hands to 9 cards for each player til we get a hand size field implemented. Only other thing worth noting is the DUPLICATE copy functionality that was added for JS was removed entirely.